feat(invoke-message): refactor message handling and introduce InvokeMessage class
