+++ python/dify_plugin/entities/agent.py
@@ -7,12 +7,12 @@ from pydantic import BaseModel, Field, field_validator
 from dify_plugin.core.documentation.schema_doc import docs
 from dify_plugin.core.utils.yaml_loader import load_yaml_file
 from dify_plugin.entities import I18nObject
+from dify_plugin.entities.invoke_message import InvokeMessage
 from dify_plugin.entities.tool import (
     CommonParameterType,
     ParameterAutoGenerate,
     ParameterTemplate,
     ToolIdentity,
-    ToolInvokeMessage,
     ToolParameterOption,
     ToolProviderIdentity,
 )
@@ -166,5 +166,5 @@ class AgentStrategyProviderConfiguration(BaseModel):
         return strategies
 
 
-class AgentInvokeMessage(ToolInvokeMessage):
+class AgentInvokeMessage(InvokeMessage):
     pass
