+++ python/dify_plugin/interfaces/tool/__init__.py
@@ -7,13 +7,14 @@ from werkzeug import Request
 
 from dify_plugin.core.runtime import Session
 from dify_plugin.entities import ParameterOption
-from dify_plugin.entities.agent import AgentInvokeMessage
-from dify_plugin.entities.tool import LogMetadata, ToolInvokeMessage, ToolParameter, ToolRuntime, ToolSelector
+from dify_plugin.entities.invoke_message import InvokeMessage
+from dify_plugin.entities.provider_config import LogMetadata
+from dify_plugin.entities.tool import ToolInvokeMessage, ToolParameter, ToolRuntime, ToolSelector
 from dify_plugin.file.constants import DIFY_FILE_IDENTITY, DIFY_TOOL_SELECTOR_IDENTITY
 from dify_plugin.file.entities import FileType
 from dify_plugin.file.file import File
 
-T = TypeVar("T", bound=ToolInvokeMessage | AgentInvokeMessage)
+T = TypeVar("T", bound=InvokeMessage)
 
 
 class ToolLike(ABC, Generic[T]):
@@ -25,14 +26,14 @@ class ToolLike(ABC, Generic[T]):
 
     def create_text_message(self, text: str) -> T:
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.TEXT,
-            message=ToolInvokeMessage.TextMessage(text=text),
+            type=InvokeMessage.MessageType.TEXT,
+            message=InvokeMessage.TextMessage(text=text),
         )
 
     def create_json_message(self, json: Mapping | list) -> T:
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.JSON,
-            message=ToolInvokeMessage.JsonMessage(json_object=json),
+            type=InvokeMessage.MessageType.JSON,
+            message=InvokeMessage.JsonMessage(json_object=json),
         )
 
     def create_image_message(self, image_url: str) -> T:
@@ -43,8 +44,8 @@ class ToolLike(ABC, Generic[T]):
         :return: the image message
         """
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.IMAGE,
-            message=ToolInvokeMessage.TextMessage(text=image_url),
+            type=InvokeMessage.MessageType.IMAGE,
+            message=InvokeMessage.TextMessage(text=image_url),
         )
 
     def create_link_message(self, link: str) -> T:
@@ -55,8 +56,8 @@ class ToolLike(ABC, Generic[T]):
         :return: the link message
         """
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.LINK,
-            message=ToolInvokeMessage.TextMessage(text=link),
+            type=InvokeMessage.MessageType.LINK,
+            message=InvokeMessage.TextMessage(text=link),
         )
 
     def create_blob_message(self, blob: bytes, meta: dict | None = None) -> T:
@@ -67,8 +68,8 @@ class ToolLike(ABC, Generic[T]):
         :return: the blob message
         """
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.BLOB,
-            message=ToolInvokeMessage.BlobMessage(blob=blob),
+            type=InvokeMessage.MessageType.BLOB,
+            message=InvokeMessage.BlobMessage(blob=blob),
             meta=meta,
         )
 
@@ -81,8 +82,8 @@ class ToolLike(ABC, Generic[T]):
         :return: the variable message
         """
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.VARIABLE,
-            message=ToolInvokeMessage.VariableMessage(variable_name=variable_name, variable_value=variable_value),
+            type=InvokeMessage.MessageType.VARIABLE,
+            message=InvokeMessage.VariableMessage(variable_name=variable_name, variable_value=variable_value),
         )
 
     def create_stream_variable_message(self, variable_name: str, variable_value: str) -> T:
@@ -96,8 +97,8 @@ class ToolLike(ABC, Generic[T]):
         :return: the variable message
         """
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.VARIABLE,
-            message=ToolInvokeMessage.VariableMessage(
+            type=InvokeMessage.MessageType.VARIABLE,
+            message=InvokeMessage.VariableMessage(
                 variable_name=variable_name,
                 variable_value=variable_value,
                 stream=True,
@@ -108,7 +109,7 @@ class ToolLike(ABC, Generic[T]):
         self,
         label: str,
         data: Mapping[str, Any],
-        status: ToolInvokeMessage.LogMessage.LogStatus = ToolInvokeMessage.LogMessage.LogStatus.SUCCESS,
+        status: InvokeMessage.LogMessage.LogStatus = InvokeMessage.LogMessage.LogStatus.SUCCESS,
         parent: T | None = None,
         metadata: Mapping[LogMetadata, Any] | None = None,
     ) -> T:
@@ -116,13 +117,13 @@ class ToolLike(ABC, Generic[T]):
         create a log message with status "start"
         """
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.LOG,
-            message=ToolInvokeMessage.LogMessage(
+            type=InvokeMessage.MessageType.LOG,
+            message=InvokeMessage.LogMessage(
                 label=label,
                 data=data,
                 status=status,
                 parent_id=parent.message.id
-                if parent and isinstance(parent.message, ToolInvokeMessage.LogMessage)
+                if parent and isinstance(parent.message, InvokeMessage.LogMessage)
                 else None,
                 metadata=metadata,
             ),
@@ -130,15 +131,15 @@ class ToolLike(ABC, Generic[T]):
 
     def create_retriever_resource_message(
         self,
-        retriever_resources: list[ToolInvokeMessage.RetrieverResourceMessage.RetrieverResource],
+        retriever_resources: list[InvokeMessage.RetrieverResourceMessage.RetrieverResource],
         context: str,
     ) -> T:
         """
         create a retriever resource message
         """
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.RETRIEVER_RESOURCES,
-            message=ToolInvokeMessage.RetrieverResourceMessage(
+            type=InvokeMessage.MessageType.RETRIEVER_RESOURCES,
+            message=InvokeMessage.RetrieverResourceMessage(
                 retriever_resources=retriever_resources,
                 context=context,
             ),
@@ -147,7 +148,7 @@ class ToolLike(ABC, Generic[T]):
     def finish_log_message(
         self,
         log: T,
-        status: ToolInvokeMessage.LogMessage.LogStatus = ToolInvokeMessage.LogMessage.LogStatus.SUCCESS,
+        status: InvokeMessage.LogMessage.LogStatus = InvokeMessage.LogMessage.LogStatus.SUCCESS,
         error: str | None = None,
         data: Mapping[str, Any] | None = None,
         metadata: Mapping[LogMetadata, Any] | None = None,
@@ -155,10 +156,10 @@ class ToolLike(ABC, Generic[T]):
         """
         mark log as finished
         """
-        assert isinstance(log.message, ToolInvokeMessage.LogMessage)
+        assert isinstance(log.message, InvokeMessage.LogMessage)
         return self.response_type(
-            type=ToolInvokeMessage.MessageType.LOG,
-            message=ToolInvokeMessage.LogMessage(
+            type=InvokeMessage.MessageType.LOG,
+            message=InvokeMessage.LogMessage(
                 id=log.message.id,
                 label=log.message.label,
                 data=data or log.message.data,
